refactor: layer 일관성을 위한 클래스 이동
