+++ src/main/java/io/github/eappezo/soundary/services/authentication/application/service/SocialUserCreationSupport.java
@@ -1,4 +1,4 @@
-package io.github.eappezo.soundary.services.authentication.domain;
+package io.github.eappezo.soundary.services.authentication.application.service;
 
 import io.github.eappezo.soundary.core.user.User;
 import io.github.eappezo.soundary.services.authentication.application.OAuthResult;
