+++ src/main/java/io/github/eappezo/soundary/services/authentication/infrastructure/JwtTokenProvider.java
@@ -3,7 +3,7 @@ package io.github.eappezo.soundary.services.authentication.infrastructure;
 import io.github.eappezo.soundary.core.identification.Identifier;
 import io.github.eappezo.soundary.core.user.User;
 import io.github.eappezo.soundary.core.user.UserRole;
-import io.github.eappezo.soundary.services.authentication.application.TokenProvider;
+import io.github.eappezo.soundary.services.authentication.domain.TokenProvider;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
 import jakarta.annotation.PostConstruct;
